image-sharpening-kd/requirements.txt
torch
torchvision
timm
scikit-image
matplotlib
numpy
Pillow
